+++ code/solution/org.gemoc.sample.legacyfsm.fsm.model/src/org/gemoc/sample/legacyfsm/fsm/impl/AssignValueActionImpl.java
@@ -32,7 +32,7 @@ public class AssignValueActionImpl extends ActionImpl implements AssignValueActi
 	 * @generated
 	 * @ordered
 	 */
-	protected static final long VALUE_EDEFAULT = 0L;
+	protected static final int VALUE_EDEFAULT = 0;
 
 	/**
 	 * The cached value of the '{@link #getValue() <em>Value</em>}' attribute.
@@ -42,7 +42,7 @@ public class AssignValueActionImpl extends ActionImpl implements AssignValueActi
 	 * @generated
 	 * @ordered
 	 */
-	protected long value = VALUE_EDEFAULT;
+	protected int value = VALUE_EDEFAULT;
 
 	/**
 	 * <!-- begin-user-doc -->
@@ -68,7 +68,7 @@ public class AssignValueActionImpl extends ActionImpl implements AssignValueActi
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public long getValue() {
+	public int getValue() {
 		return value;
 	}
 
@@ -77,8 +77,8 @@ public class AssignValueActionImpl extends ActionImpl implements AssignValueActi
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public void setValue(long newValue) {
-		long oldValue = value;
+	public void setValue(int newValue) {
+		int oldValue = value;
 		value = newValue;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET, FsmPackage.ASSIGN_VALUE_ACTION__VALUE, oldValue, value));
@@ -107,7 +107,7 @@ public class AssignValueActionImpl extends ActionImpl implements AssignValueActi
 	public void eSet(int featureID, Object newValue) {
 		switch (featureID) {
 			case FsmPackage.ASSIGN_VALUE_ACTION__VALUE:
-				setValue((Long)newValue);
+				setValue((Integer)newValue);
 				return;
 		}
 		super.eSet(featureID, newValue);
