+++ code/solution/org.gemoc.sample.legacyfsm.xsfsm/src-gen/org/gemoc/sample/legacyfsm/xsfsm/xsfsm/adapters/xsfsmmt/fsm/GreaterThanNumberGuardAdapter.java
@@ -26,12 +26,12 @@ public class GreaterThanNumberGuardAdapter extends EObjectAdapter<GreaterThanNum
   }
   
   @Override
-  public long getValue() {
+  public int getValue() {
     return adaptee.getValue();
   }
   
   @Override
-  public void setValue(final long o) {
+  public void setValue(final int o) {
     adaptee.setValue(o);
   }
   
@@ -54,7 +54,7 @@ public class GreaterThanNumberGuardAdapter extends EObjectAdapter<GreaterThanNum
   
   protected final static boolean NOT_EDEFAULT = false;
   
-  protected final static long VALUE_EDEFAULT = -1L;
+  protected final static int VALUE_EDEFAULT = -1;
   
   @Override
   public EClass eClass() {
@@ -67,7 +67,7 @@ public class GreaterThanNumberGuardAdapter extends EObjectAdapter<GreaterThanNum
     	case org.gemoc.sample.legacyfsm.xsfsm.xsfsmmt.fsm.FsmPackage.GREATER_THAN_NUMBER_GUARD__NOT:
     		return isNot() ? Boolean.TRUE : Boolean.FALSE;
     	case org.gemoc.sample.legacyfsm.xsfsm.xsfsmmt.fsm.FsmPackage.GREATER_THAN_NUMBER_GUARD__VALUE:
-    		return new java.lang.Long(getValue());
+    		return new java.lang.Integer(getValue());
     	case org.gemoc.sample.legacyfsm.xsfsm.xsfsmmt.fsm.FsmPackage.GREATER_THAN_NUMBER_GUARD__SOURCE:
     		return getSource();
     }
@@ -96,7 +96,7 @@ public class GreaterThanNumberGuardAdapter extends EObjectAdapter<GreaterThanNum
     		setNot(((java.lang.Boolean) newValue).booleanValue());
     		return;
     	case org.gemoc.sample.legacyfsm.xsfsm.xsfsmmt.fsm.FsmPackage.GREATER_THAN_NUMBER_GUARD__VALUE:
-    		setValue(((java.lang.Long) newValue).longValue());
+    		setValue(((java.lang.Integer) newValue).intValue());
     		return;
     	case org.gemoc.sample.legacyfsm.xsfsm.xsfsmmt.fsm.FsmPackage.GREATER_THAN_NUMBER_GUARD__SOURCE:
     		setSource(
