+++ code/solution/org.gemoc.sample.legacyfsm.xsfsm/src-gen/org/gemoc/sample/legacyfsm/xsfsm/xsfsm/adapters/xsfsmmt/fsm/IncreaseValueActionAdapter.java
@@ -16,12 +16,12 @@ public class IncreaseValueActionAdapter extends EObjectAdapter<IncreaseValueActi
   }
   
   @Override
-  public long getStepValue() {
+  public int getStepValue() {
     return adaptee.getStepValue();
   }
   
   @Override
-  public void setStepValue(final long o) {
+  public void setStepValue(final int o) {
     adaptee.setStepValue(o);
   }
   
@@ -42,7 +42,7 @@ public class IncreaseValueActionAdapter extends EObjectAdapter<IncreaseValueActi
     org.gemoc.sample.legacyfsm.xsfsm.xsfsm.aspects.IncreaseValueActionAspect.execute(adaptee);
   }
   
-  protected final static long STEP_VALUE_EDEFAULT = 0L;
+  protected final static int STEP_VALUE_EDEFAULT = 0;
   
   @Override
   public EClass eClass() {
@@ -55,7 +55,7 @@ public class IncreaseValueActionAdapter extends EObjectAdapter<IncreaseValueActi
     	case org.gemoc.sample.legacyfsm.xsfsm.xsfsmmt.fsm.FsmPackage.INCREASE_VALUE_ACTION__TARGET:
     		return getTarget();
     	case org.gemoc.sample.legacyfsm.xsfsm.xsfsmmt.fsm.FsmPackage.INCREASE_VALUE_ACTION__STEP_VALUE:
-    		return new java.lang.Long(getStepValue());
+    		return new java.lang.Integer(getStepValue());
     }
     
     return super.eGet(featureID, resolve, coreType);
@@ -82,7 +82,7 @@ public class IncreaseValueActionAdapter extends EObjectAdapter<IncreaseValueActi
     		 newValue);
     		return;
     	case org.gemoc.sample.legacyfsm.xsfsm.xsfsmmt.fsm.FsmPackage.INCREASE_VALUE_ACTION__STEP_VALUE:
-    		setStepValue(((java.lang.Long) newValue).longValue());
+    		setStepValue(((java.lang.Integer) newValue).intValue());
     		return;
     }
